Update README.md
The goals of this project is to track the status of patient wait list for an ER room, analyze the historical monthly wait room trends for both inpatient and outpatient, as well as identify specialty level with age profile analysis. The scope will be from 2018-2021, metrics needed are average and median waiting list along with current total wait list. Summary page and detailed page is structured for viewing and analysis.